requirements.txt
lxml==4.4.1
pkg-resources==0.0.0
py==1.8.0
pyshark==0.4.2.9
python-statemachine==0.7.1
